requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
google-cloud-bigquery==3.19.0
pytest==8.1.1
pytest-cov==4.1.0